Dedicated macros for memory region start/end
