+++ esp-hal/Cargo.toml
@@ -39,6 +39,7 @@ embedded-io-async        = { version = "0.6.1", optional = true }
 enumset                  = "1.1.5"
 esp-build                = { version = "0.2.0", path = "../esp-build" }
 esp-config               = { version = "0.3.0", path = "../esp-config" }
+esp-metadata             = { version = "0.6.0", path = "../esp-metadata", default-features = false }
 esp-synopsys-usb-otg     = { version = "0.4.2", optional = true, features = ["fs", "esp32sx"] }
 fugit                    = "0.3.7"
 instability              = "0.3.7"
